Throw away all we've done and start making controls that are in line with the vision (think Gears of Ware 3rd person controls)
