+++ .vs/SooOOOMO/xs/UserPrefs.xml
@@ -1,13 +1,15 @@
 ﻿<Properties StartupConfiguration="{7BB6A468-2748-E70A-E492-B15E5D3EF757}|">
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Programming/PersistAcrossScenes.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Programming/RikishiController.cs">
     <Files>
       <File FileName="Assets/Programming/FadeToBlack.cs" Line="25" Column="6" />
-      <File FileName="Assets/Programming/RikishiController.cs" Line="27" Column="53" />
       <File FileName="Assets/Programming/ResetGameIfTouched.cs" Line="16" Column="6" />
       <File FileName="Assets/Programming/Scoreboard.cs" Line="37" Column="12" />
-      <File FileName="Assets/Programming/DetectEnemyAlertRikishi.cs" Line="30" Column="48" />
-      <File FileName="Assets/Programming/GameController.cs" Line="20" Column="18" />
-      <File FileName="Assets/Programming/PersistAcrossScenes.cs" Line="19" Column="1" />
+      <File FileName="Assets/Programming/GameController.cs" Line="25" Column="2" />
+      <File FileName="Assets/Programming/PersistAcrossScenes.cs" Line="18" Column="2" />
+      <File FileName="Assets/Standard Assets/Characters/ThirdPersonCharacter/Scripts/ThirdPersonCharacter.cs" Line="121" Column="23" />
+      <File FileName="Assets/Standard Assets/Characters/ThirdPersonCharacter/Scripts/ThirdPersonUserControl.cs" Line="71" Column="28" />
+      <File FileName="Assets/Programming/RikishiProtoController.cs" Line="10" Column="17" />
+      <File FileName="Assets/Programming/RikishiController.cs" Line="14" Column="3" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
