+++ Assets/Animation/RikishiAnimator.controller
@@ -274,7 +274,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400002, guid: f8a71f54f1a1a48f6a573c7acfb5a087, type: 3}
+  m_Motion: {fileID: 7400006, guid: 973312c4ec397464ea682665e64638e2, type: 3}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -300,7 +300,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400016, guid: 973312c4ec397464ea682665e64638e2, type: 3}
+  m_Motion: {fileID: 7400018, guid: 973312c4ec397464ea682665e64638e2, type: 3}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -361,8 +361,8 @@ AnimatorStateMachine:
   m_EntryTransitions: []
   m_StateMachineTransitions: {}
   m_StateMachineBehaviours: []
-  m_AnyStatePosition: {x: 96, y: 48, z: 0}
-  m_EntryPosition: {x: 96, y: 132, z: 0}
+  m_AnyStatePosition: {x: 96, y: -24, z: 0}
+  m_EntryPosition: {x: 96, y: 48, z: 0}
   m_ExitPosition: {x: 96, y: 240, z: 0}
   m_ParentStateMachinePosition: {x: 800, y: 20, z: 0}
-  m_DefaultState: {fileID: 1102240036293543474}
+  m_DefaultState: {fileID: 1102035140914799232}
