--- Assets/Models/Rikishi_v10_modeled-ish-rigged-ish-animated-ish-SONKYO.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 6c8d0e0be02c643549bec39b6f6faf06
-timeCreated: 1530941163
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
