--- Assets/Models/Rikishi_v2_wip.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 013a2308516d74314b65c37907e8f39c
-timeCreated: 1530585319
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
