--- Assets/Models/Rikishi_v4_modeled-ish.blend.meta
@@ -1,146 +1,0 @@
-fileFormatVersion: 2
-guid: e31ecec86f35b4505a248f672adadb77
-timeCreated: 1530585322
-licenseType: Free
-ModelImporter:
-  serializedVersion: 22
-  fileIDToRecycleName:
-    100000: //RootNode
-    100002: Armature
-    100004: Cube
-    100006: Left_Clavicle
-    100008: Left_Foot
-    100010: Left_Hand
-    100012: Left_Hip
-    100014: Left_Lower_Arm
-    100016: Left_Lower_Leg
-    100018: Left_Shoulder
-    100020: Left_Upper_Arm
-    100022: Left_Upper_Leg
-    100024: Main
-    100026: Middle_Spine
-    100028: Neck_Main
-    100030: Neck_Sub
-    100032: Right_Clavicle
-    100034: Right_Foot
-    100036: Right_Hand
-    100038: Right_Hip
-    100040: Right_Lower_Arm
-    100042: Right_Lower_Leg
-    100044: Right_Shoulder
-    100046: Right_Upper_Arm
-    100048: Right_Upper_Leg
-    100050: Sumo_Skull
-    400000: //RootNode
-    400002: Armature
-    400004: Cube
-    400006: Left_Clavicle
-    400008: Left_Foot
-    400010: Left_Hand
-    400012: Left_Hip
-    400014: Left_Lower_Arm
-    400016: Left_Lower_Leg
-    400018: Left_Shoulder
-    400020: Left_Upper_Arm
-    400022: Left_Upper_Leg
-    400024: Main
-    400026: Middle_Spine
-    400028: Neck_Main
-    400030: Neck_Sub
-    400032: Right_Clavicle
-    400034: Right_Foot
-    400036: Right_Hand
-    400038: Right_Hip
-    400040: Right_Lower_Arm
-    400042: Right_Lower_Leg
-    400044: Right_Shoulder
-    400046: Right_Upper_Arm
-    400048: Right_Upper_Leg
-    400050: Sumo_Skull
-    2100000: unnamed
-    2300000: //RootNode
-    3300000: //RootNode
-    4300000: Cube
-    7400000: Default Take
-    9500000: //RootNode
-    13700000: Cube
-  externalObjects: {}
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-    materialLocation: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 0
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    importAnimatedCustomProperties: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    extraUserProperties: []
-    clipAnimations: []
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importVisibility: 1
-    importBlendShapes: 1
-    importCameras: 1
-    importLights: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    preserveHierarchy: 0
-    indexFormat: 0
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 1
-    tangentImportMode: 3
-    normalCalculationMode: 4
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: 
-    rootMotionBoneRotation: {x: 0, y: 0, z: 0, w: 1}
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
