--- Assets/Models/Rikishi_v4_modeled-ish.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: db73ec0f731634a458d757d65f7d232f
-timeCreated: 1530587822
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
