--- Assets/Models/Rikishi_v5_modeled-ish-rigged-ish.blend.meta
@@ -1,144 +1,0 @@
-fileFormatVersion: 2
-guid: 1723b227cf3e845819ba4cc31a3b71a7
-timeCreated: 1530590457
-licenseType: Free
-ModelImporter:
-  serializedVersion: 22
-  fileIDToRecycleName:
-    100000: Armature
-    100002: Cube
-    100004: Left_Clavicle
-    100006: Left_Foot
-    100008: Left_Hand
-    100010: Left_Hip
-    100012: Left_Lower_Arm
-    100014: Left_Lower_Leg
-    100016: Left_Shoulder
-    100018: Left_Upper_Arm
-    100020: Left_Upper_Leg
-    100022: Main
-    100024: Middle_Spine
-    100026: Neck_Main
-    100028: Neck_Sub
-    100030: Right_Clavicle
-    100032: Right_Foot
-    100034: Right_Hand
-    100036: Right_Hip
-    100038: Right_Lower_Arm
-    100040: Right_Lower_Leg
-    100042: Right_Shoulder
-    100044: Right_Upper_Arm
-    100046: Right_Upper_Leg
-    100048: //RootNode
-    100050: Sumo_Skull
-    400000: Armature
-    400002: Cube
-    400004: Left_Clavicle
-    400006: Left_Foot
-    400008: Left_Hand
-    400010: Left_Hip
-    400012: Left_Lower_Arm
-    400014: Left_Lower_Leg
-    400016: Left_Shoulder
-    400018: Left_Upper_Arm
-    400020: Left_Upper_Leg
-    400022: Main
-    400024: Middle_Spine
-    400026: Neck_Main
-    400028: Neck_Sub
-    400030: Right_Clavicle
-    400032: Right_Foot
-    400034: Right_Hand
-    400036: Right_Hip
-    400038: Right_Lower_Arm
-    400040: Right_Lower_Leg
-    400042: Right_Shoulder
-    400044: Right_Upper_Arm
-    400046: Right_Upper_Leg
-    400048: //RootNode
-    400050: Sumo_Skull
-    2100000: unnamed
-    4300000: Cube
-    7400000: Default Take
-    9500000: //RootNode
-    13700000: Cube
-  externalObjects: {}
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-    materialLocation: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 0
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    importAnimatedCustomProperties: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    extraUserProperties: []
-    clipAnimations: []
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importVisibility: 1
-    importBlendShapes: 1
-    importCameras: 1
-    importLights: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    preserveHierarchy: 0
-    indexFormat: 0
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 0
-    tangentImportMode: 3
-    normalCalculationMode: 4
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: 
-    rootMotionBoneRotation: {x: 0, y: 0, z: 0, w: 1}
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
