--- Assets/Models/Rikishi_v5_modeled-ish-rigged-ish.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 951faa502ae044acca64dc86310120e9
-timeCreated: 1530748759
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
