--- Assets/Models/Rikishi_v8_modeled-ish-rigged-ish-animated-ish-SONKYO.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 9dda1734c365d4607b16761ff32a4888
-timeCreated: 1530760404
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
