--- Assets/Models/Rikishi_v9_modeled-ish-rigged-ish-animated-ish-SONKYO.blend1.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ab2c6eb9746de47d99b828ac2d1e22fb
-timeCreated: 1530938681
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
