+++ Assets/Prefabs/Rikishi_proto.prefab.meta
