+++ Assets/Programming/RikishiProtoController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5ef1028e5a7814d34bb2340c2028d046
-timeCreated: 1531271242
+guid: 1b84cc346be4840cab87fc2e2068e777
+timeCreated: 1530939002
 licenseType: Free
 MonoImporter:
   externalObjects: {}
